Chunked downloads

+++ src/CyberdropDownloader.Core/AlbumDownloader.cs
@@ -1,9 +1,13 @@
 ﻿using CyberdropDownloader.Core.DataModels;
 using System;
+using System.Collections.Generic;
 using System.IO;
+using System.Net;
 using System.Net.Http;
+using System.Net.Http.Headers;
 using System.Text.RegularExpressions;
 using System.Threading;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace CyberdropDownloader.Core
@@ -14,6 +18,9 @@ namespace CyberdropDownloader.Core
         private bool _authorized;
         private bool _running;
 
+        private List<Chunk> _chunks;
+        private Chunk _currentChunk;
+
         public AlbumDownloader(bool authorized)
         {
             // Setup and initialize HttpClient
@@ -23,8 +30,11 @@ namespace CyberdropDownloader.Core
                 KeepAlivePingTimeout = Timeout.InfiniteTimeSpan
             });
 
-            // Times out current download if it surpasses 30 minutes (might need tweaked)
-            _downloadClient.Timeout = TimeSpan.FromMinutes(30);
+            // Times out current download if it surpasses 5 minutes (might need tweaked)
+            _downloadClient.Timeout = TimeSpan.FromMinutes(5);
+
+            ServicePointManager.DefaultConnectionLimit = 100;
+            ServicePointManager.Expect100Continue = false;
 
             _authorized = authorized;
         }
@@ -33,14 +43,13 @@ namespace CyberdropDownloader.Core
         public bool Authorized { get => _authorized; set => _authorized = value; }
         public bool Running { get => _running; }
 
-        public delegate void EventHandler(string fileName);
+        public event EventHandler<string> FileDownloaded;
+        public event EventHandler<string> FileDownloading;
+        public event EventHandler<string> FileFailed;
+        public event EventHandler<string> FileExists;
+        public event EventHandler<int> ProgressChanged;
 
-        public event EventHandler FileDownloaded;
-        public event EventHandler FileDownloading;
-        public event EventHandler FileFailed;
-        public event EventHandler FileExists;
-
-        public async Task DownloadAsync(Album album, string path, CancellationTokenSource cancellationTokenSource)
+        public async Task DownloadAsync(Album album, string path, CancellationToken? cancellationToken, int chunkCount = 1)
         {
             // If not authorized throw exception
             if (!_authorized)
@@ -55,8 +64,8 @@ namespace CyberdropDownloader.Core
             // Iterate through all album files until there are none left
             while (album.Files.Count > 0)
             {
-                cancellationTokenSource.Token.ThrowIfCancellationRequested();
-                
+                cancellationToken.Value.ThrowIfCancellationRequested();
+
                 _running = true;
 
                 AlbumFile file = album.Files.Peek();
@@ -65,52 +74,92 @@ namespace CyberdropDownloader.Core
 
                 try
                 {
-                    FileDownloading.Invoke(file.Name);
-
                     HttpResponseMessage response = await _downloadClient.GetAsync(file.Url, HttpCompletionOption.ResponseHeadersRead);
+
+                    if (response.ReasonPhrase == "Bad Gateway")
+                        throw new Exception(response.ReasonPhrase);
 
                     while (response.ReasonPhrase == "Moved Temporarily")
                         response = await _downloadClient.GetAsync(response.Headers.Location, HttpCompletionOption.ResponseHeadersRead);
 
-                    using (Stream dataStream = await response.Content.ReadAsStreamAsync(cancellationTokenSource.Token))
+                    if (File.Exists(filePath))
                     {
-                        if (File.Exists(filePath))
+                        long fileLength = new FileInfo(filePath).Length;
+
+                        if (fileLength == response.Content.Headers.ContentLength)
                         {
-                            long fileLength = new FileInfo(filePath).Length;
-
-                            if (fileLength == response.Content.Headers.ContentLength)
-                            {
-                                FileExists.Invoke(album.Files.Dequeue().Name);
-                                _running = false;
-                                continue;
-                            }
-                            else File.Delete(filePath);
+                            FileExists?.Invoke(this, album.Files.Dequeue().Name);
+                            _running = false;
+                            continue;
                         }
-
-                        using (Stream fileStream = File.Open(filePath, FileMode.Create, FileAccess.Write))
-                        {
-                            await dataStream.CopyToAsync(fileStream, cancellationTokenSource.Token);
-                        }
-
-                        response.Dispose();
-
-                        FileDownloaded.Invoke(album.Files.Dequeue().Name);
+                        else File.Delete(filePath);
                     }
+
+                    _chunks = new List<Chunk>();
+
+                    for (int chunk = 0; chunk <= chunkCount; chunk++)
+                    {
+                        _chunks.Add(new Chunk()
+                        {
+                            Start = chunk * (response.Content.Headers.ContentLength.Value / chunkCount),
+                            End = (chunk + 1) * (response.Content.Headers.ContentLength.Value / chunkCount)
+                        });
+                    }
+
+                    _chunks.LastOrDefault().End = response.Content.Headers.ContentLength.Value;
+
+                    //FileStream fileStream = File.Open(filePath, FileMode.Append, FileAccess.Write, FileShare.None);
+                    FileStream fileStream = File.OpenWrite(filePath);
+
+                    FileDownloading?.Invoke(this, file.Name);
+
+                    while (_chunks.Count > 0)
+                    {
+                        cancellationToken.Value.ThrowIfCancellationRequested();
+
+                        _currentChunk = _chunks[0];
+
+                        try
+                        {
+                            using (HttpRequestMessage request = new HttpRequestMessage())
+                            {
+                                request.RequestUri = new Uri(file.Url);
+                                request.Headers.Range = new RangeHeaderValue(_currentChunk.Start, _currentChunk.End);
+
+                                using (HttpResponseMessage rangedResponse = await _downloadClient.SendAsync(request, cancellationToken.Value))
+                                {
+                                    fileStream.Seek(_currentChunk.Start, SeekOrigin.Begin);
+                                    await fileStream.WriteAsync(await rangedResponse.Content.ReadAsByteArrayAsync(), cancellationToken.Value);
+                                }
+                            }
+                        }
+                        catch (Exception) { continue; }
+
+                        _chunks.RemoveAt(0);
+                        ProgressChanged?.Invoke(this, chunkCount - _chunks.Count);
+                    }
+
+                    FileDownloaded?.Invoke(this, album.Files.Dequeue().Name);
                 }
-                catch (Exception) { FileFailed.Invoke(file.Name); }
+                catch (Exception ex) 
+                {
+                    string message = ex.Message;
+
+                    FileFailed?.Invoke(this, album.Files.Dequeue().Name); 
+                }
 
                 _running = false;
             }
         }
 
-        public void CancelDownload() 
+        public void CancelDownload()
         {
             // Cancels pending get requests
             _downloadClient.CancelPendingRequests();
 
             // Indicates that no file i currently being downloaded
             _running = false;
-        } 
+        }
 
         private static string NormalizeFileName(string data)
         {
